using c++ casts instead of c casts
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1066 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_dyngrid_layout.cpp
@@ -371,11 +371,11 @@ QList<QRect> QwtDynGridLayout::layoutItems( const QRect &rect,
     const int xySpace = spacing();
 
     rowY[0] = yOffset + margin();
-    for ( int r = 1; r < ( int )numRows; r++ )
+    for ( uint r = 1; r < numRows; r++ )
         rowY[r] = rowY[r-1] + rowHeight[r-1] + xySpace;
 
     colX[0] = xOffset + margin();
-    for ( int c = 1; c < ( int )numCols; c++ )
+    for ( uint c = 1; c < numCols; c++ )
         colX[c] = colX[c-1] + colWidth[c-1] + xySpace;
 
     const int itemCount = d_data->itemList.size();
@@ -411,8 +411,7 @@ void QwtDynGridLayout::layoutGrid( uint numCols,
     if ( d_data->isDirty )
         d_data->updateLayoutCache();
 
-    for ( uint index = 0;
-        index < ( uint )d_data->itemSizeHints.count(); index++ )
+    for ( int index = 0; index < d_data->itemSizeHints.count(); index++ )
     {
         const int row = index / numCols;
         const int col = index % numCols;
@@ -455,7 +454,7 @@ int QwtDynGridLayout::heightForWidth( int width ) const
     layoutGrid( numCols, rowHeight, colWidth );
 
     int h = 2 * margin() + ( numRows - 1 ) * spacing();
-    for ( int row = 0; row < ( int )numRows; row++ )
+    for ( uint row = 0; row < numRows; row++ )
         h += rowHeight[row];
 
     return h;
@@ -481,12 +480,12 @@ void QwtDynGridLayout::stretchGrid( const QRect &rect,
     if ( expandH )
     {
         int xDelta = rect.width() - 2 * margin() - ( numCols - 1 ) * spacing();
-        for ( int col = 0; col < ( int )numCols; col++ )
+        for ( uint col = 0; col < numCols; col++ )
             xDelta -= colWidth[col];
 
         if ( xDelta > 0 )
         {
-            for ( int col = 0; col < ( int )numCols; col++ )
+            for ( uint col = 0; col < numCols; col++ )
             {
                 const int space = xDelta / ( numCols - col );
                 colWidth[col] += space;
@@ -502,12 +501,12 @@ void QwtDynGridLayout::stretchGrid( const QRect &rect,
             numRows++;
 
         int yDelta = rect.height() - 2 * margin() - ( numRows - 1 ) * spacing();
-        for ( int row = 0; row < ( int )numRows; row++ )
+        for ( uint row = 0; row < numRows; row++ )
             yDelta -= rowHeight[row];
 
         if ( yDelta > 0 )
         {
-            for ( int row = 0; row < ( int )numRows; row++ )
+            for ( uint row = 0; row < numRows; row++ )
             {
                 const int space = yDelta / ( numRows - row );
                 rowHeight[row] += space;
@@ -540,11 +539,11 @@ QSize QwtDynGridLayout::sizeHint() const
     layoutGrid( numCols, rowHeight, colWidth );
 
     int h = 2 * margin() + ( numRows - 1 ) * spacing();
-    for ( int row = 0; row < ( int )numRows; row++ )
+    for ( uint row = 0; row < numRows; row++ )
         h += rowHeight[row];
 
     int w = 2 * margin() + ( numCols - 1 ) * spacing();
-    for ( int col = 0; col < ( int )numCols; col++ )
+    for ( uint col = 0; col < numCols; col++ )
         w += colWidth[col];
 
     return QSize( w, h );
